+++ src/main/java/pl/mantiscrab/budgetr/config/SecurityConfig.java
@@ -3,6 +3,7 @@ package pl.mantiscrab.budgetr.config;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
@@ -15,8 +16,10 @@ import javax.sql.DataSource;
 @Configuration
 class SecurityConfig {
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+    @Profile("dev")
+    public SecurityFilterChain filterChainDev(HttpSecurity http) throws Exception {
         http.authorizeHttpRequests(requests -> requests
+                .mvcMatchers("/register").permitAll()
                 .mvcMatchers("/v3/api-docs/**", "/swagger-ui/**", "/swagger-ui.html").permitAll()
                 .requestMatchers(PathRequest.toH2Console()).permitAll()
                 .anyRequest().authenticated());
@@ -24,6 +27,18 @@ class SecurityConfig {
         http.httpBasic();
         http.csrf().disable();
         http.headers().frameOptions().sameOrigin();
+        return http.build();
+    }
+
+    @Bean
+    @Profile("test")
+    public SecurityFilterChain filterChainTest(HttpSecurity http) throws Exception {
+        http.authorizeHttpRequests(requests -> requests
+                .mvcMatchers("/register").permitAll()
+                .anyRequest().authenticated());
+        http.formLogin();
+        http.httpBasic();
+        http.csrf().disable();
         return http.build();
     }
